# apps/api/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WMS API",
    "version": "1.0.0",
    "description": "Minimal WMS API with Swagger UI"
  },
  "servers": [
    {
      "url": "http://localhost:3100"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/clients": {
      "get": {
        "summary": "List clients",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Create client",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "client_code",
                  "name_kr"
                ],
                "properties": {
                  "client_code": {
                    "type": "string"
                  },
                  "name_kr": {
                    "type": "string"
                  },
                  "name_en": {
                    "type": "string"
                  },
                  "contact_name": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "example": "active"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "409": {
            "description": "Duplicate client_code"
          }
        }
      }
    },
    "/clients/{id}": {
      "get": {
        "summary": "Get client",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "Update client",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "client_code",
                  "name_kr",
                  "status"
                ],
                "properties": {
                  "client_code": {
                    "type": "string"
                  },
                  "name_kr": {
                    "type": "string"
                  },
                  "name_en": {
                    "type": "string"
                  },
                  "contact_name": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Delete client (soft)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/inbound-items": {
      "get": {
        "summary": "List inbound items",
        "parameters": [
          {
            "name": "inbound_order_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Create inbound item",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "inbound_order_id",
                  "product_id",
                  "lot_id",
                  "qty"
                ],
                "properties": {
                  "inbound_order_id": {
                    "type": "integer"
                  },
                  "product_id": {
                    "type": "integer"
                  },
                  "lot_id": {
                    "type": "integer"
                  },
                  "location_id": {
                    "type": "integer"
                  },
                  "qty": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "invoice_price": {
                    "type": "number"
                  },
                  "currency": {
                    "type": "string",
                    "example": "THB"
                  },
                  "remark": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation/FK error"
          }
        }
      }
    },
    "/inbound-items/{id}": {
      "get": {
        "summary": "Get inbound item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "Update inbound item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "inbound_order_id",
                  "product_id",
                  "lot_id",
                  "qty"
                ],
                "properties": {
                  "inbound_order_id": {
                    "type": "integer"
                  },
                  "product_id": {
                    "type": "integer"
                  },
                  "lot_id": {
                    "type": "integer"
                  },
                  "location_id": {
                    "type": "integer"
                  },
                  "qty": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "invoice_price": {
                    "type": "number"
                  },
                  "currency": {
                    "type": "string"
                  },
                  "remark": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Delete inbound item (soft)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  }
}